pyproject.toml
[tool.poetry]
name = "MiningNiti"
version = "0.1.0"
description = "MiningNiti"
authors = ["Pulkit077 <pulkitagarwal2k1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.3.0"
langchain = "^0.0.350"
pyyaml = "^6.0.1"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
streamlit = "^1.29.0"
langchain-google-genai = "^0.0.2"
speechrecognition = "^3.10.1"
gtts = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
